+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 ### Fixes
 - Dialog 2D scale not fixed based on gui scale
 - Entity played from a playback during subscene recording not ignored
+- Can't play, stop or change locale to more than one player using command